requirements.txt
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
streamlit==1.38.0
yfinance==0.2.44
python-dotenv==0.21.0